inline server
