+++ examples/python/auto-tracing/app/templates/index.html
