+++ examples/python/auto-tracing/main.py
@@ -3,7 +3,7 @@ import logfire
 logfire.configure()
 logfire.install_auto_tracing(modules=['app'], min_duration=0)
 
-from app.server import app, db  # noqa
+from app import app, db  # noqa  # needs to be imported after auto-tracing
 
 db.init_app(app)
